About the issue #10, the progressBar is represented by a decimal number between 0 and 1.

+++ ComicsShelf/Database/ComicFiles.cs
@@ -19,7 +19,7 @@ namespace ComicsShelf.Database
       public short? Rate { get; set; }
 
       public bool Readed { get; set; }
-      public short ReadingPercent { get; set; }
+      public double ReadingPercent { get; set; }
       public short ReadingPage { get; set; }
       public string ReadingDate { get; set; }
